Refactor: move placeholder word breaker to word_breakers directory/namespace.
